Moved the cljs-lib compiled files from cljs-out to out/clsj-lib.
